requirements.txt
fastapi
uvicorn
motor
python-dotenv
jinja2
